// real-estate-crm-dashboard-react/src/data/generated/katalystDocuments.ts
export const katalystDocuments = {
  "byNormalizedAddress": {
    "1137 26th st des moines ia": [
      {
        "name": "Offering Memorandum",
        "url": "https://katalystteam.com/wp-content/uploads/2026/04/1137-26th-St.-Des-Moines-IA_OM_4.20.2026_compressed.pdf"
      },
      {
        "name": "Income Statements 2025",
        "url": "https://katalystteam.com/wp-content/uploads/2026/04/1137-26th-St.-Des-Moines-IA_Income-Statement-2025_4.21.2026.xlsx"
      },
      {
        "name": "Income Statements 2024",
        "url": "https://katalystteam.com/wp-content/uploads/2026/04/1137-26th-St.-Des-Moines-IA_Income-Statement-2024_4.21.2026.xlsx"
      }
    ],
    "2733-2897 boston ave des moines ia": [
      {
        "name": "Offering Memorandum",
        "url": "https://katalystteam.com/wp-content/uploads/2026/03/2733-2897-Boston-Ave-OM-c2.pdf"
      },
      {
        "name": "Rent Roll",
        "url": "https://katalystteam.com/wp-content/uploads/2026/03/2733-2897-Boston-Ave.-Des-Moines-IA_Rent-Roll.xlsx"
      }
    ],
    "200 e arlington st waterloo ia": [
      {
        "name": "Offering Memorandum",
        "url": "https://katalystteam.com/wp-content/uploads/2025/12/OM_200_E_Arlington_St_c2.pdf"
      },
      {
        "name": "4 Pager",
        "url": "https://katalystteam.com/wp-content/uploads/2025/12/4-page-Flyer_200-E.-Arlington-St.-Waterloo-IA.pdf"
      }
    ],
    "811 16th st des moines ia": [
      {
        "name": "Offering Memorandum",
        "url": "https://katalystteam.com/wp-content/uploads/2025/11/OM_811-16th-St_Des-Moines_IA_c2.pdf"
      },
      {
        "name": "4 Pager",
        "url": "https://katalystteam.com/wp-content/uploads/2025/11/4-Pager_811-16th-St_Des-Moines_IA.pdf"
      },
      {
        "name": "Rent Roll",
        "url": "https://katalystteam.com/wp-content/uploads/2026/01/811-16th-St-Des-Moines-IA_Rent-Roll.xlsx"
      },
      {
        "name": "Income Statement",
        "url": "https://katalystteam.com/wp-content/uploads/2026/01/811-16th-St-Des-Moines-IA_Income-Statement-1125-to-123125.xlsx"
      }
    ],
    "2411 welbeck rd des moines ia": [
      {
        "name": "Offering Memorandum",
        "url": "https://katalystteam.com/wp-content/uploads/2025/09/OM_2_2411-Welbeck-Rd.pdf"
      },
      {
        "name": "4 Pager",
        "url": "https://katalystteam.com/wp-content/uploads/2025/09/4-pager_2411-Welbeck-Rd.pdf"
      },
      {
        "name": "Rent Roll",
        "url": "https://katalystteam.com/wp-content/uploads/2026/01/2411-Welbeck-Rd-Des-Moines-IA_Rent-Roll.pdf"
      },
      {
        "name": "Cash Flow",
        "url": "https://katalystteam.com/wp-content/uploads/2026/01/2411-Welbeck-Rd-Des-Moines-IA_Cash-Flow-2025.pdf"
      }
    ],
    "301 se 8th st des moines ia": [
      {
        "name": "Offering Memorandum",
        "url": "https://katalystteam.com/wp-content/uploads/2025/09/OM_301-309-SE-8th-St-Des-Moines-IA-50390.pdf"
      },
      {
        "name": "4 Pager",
        "url": "https://katalystteam.com/wp-content/uploads/2025/09/4-Page-Flyer_301-309-SE-8th-St-Des-Moines-IA-50390.pdf"
      }
    ],
    "1805 arlington ave des moines ia": [
      {
        "name": "Offering Memorandum",
        "url": "https://katalystteam.com/wp-content/uploads/2025/03/1805-Arlington-Avenue-OM-Compressed.pdf"
      },
      {
        "name": "Pro-Forma",
        "url": "https://katalystteam.com/wp-content/uploads/2026/01/1805-Arlington-Ave-Des-Moines-IA_Pro-Forma.xlsx"
      },
      {
        "name": "Rent Roll",
        "url": "https://katalystteam.com/wp-content/uploads/2026/01/1805-Arlington-Ave-Des-Moines-IA_Rent-Roll-as-of-12125_2725.xlsx"
      }
    ],
    "405 n davis st anamosa ia": [
      {
        "name": "4 Pager",
        "url": "https://katalystteam.com/wp-content/uploads/2025/10/FINAL_405_N_Davis_St_4_Pager.pdf"
      },
      {
        "name": "Offering Memorandum",
        "url": "https://katalystteam.com/wp-content/uploads/2025/10/FINAL_405_N_Davis_St_OM_Comp2.pdf"
      },
      {
        "name": "T12",
        "url": "https://katalystteam.com/wp-content/uploads/2026/01/405-N-Davis-St-Anamosa-IA_T12_8-23-24.pdf"
      },
      {
        "name": "Rent Roll",
        "url": "https://katalystteam.com/wp-content/uploads/2026/01/405-N-Davis-St-Anamosa-IA_RR-7-30-24.xlsx"
      },
      {
        "name": "YTD",
        "url": "https://katalystteam.com/wp-content/uploads/2026/01/405-N-Davis-YTD-9302025.pdf"
      }
    ],
    "3703 3707 hubbell ave des moines ia": [
      {
        "name": "4 Pager",
        "url": "https://katalystteam.com/wp-content/uploads/2024/04/3703-Hubbell-Ave_Digital.pdf"
      }
    ]
  }
} as const;