// real-estate-crm-dashboard-react/src/data/generated/gdriveDocuments.ts
/* eslint-disable */
// Auto-generated by `npm run extract:gdrive-documents`. Maps listing addresses to Google Drive folders.
export const gdriveDocuments = {
  "rootFolderId": "1LMtw2fnyYvY8NWKxrZVSDVRzhgO-dvDc",
  "byNormalizedAddress": {
    "50 broadway st central city ia": "https://drive.google.com/drive/folders/1oN3lOnmtccHMhEqxYWT7Dk2rjMuF8W5C",
    "triangle court osage ia - mobile home park": "https://drive.google.com/drive/folders/12MJFxSau4c35YXD_rZzLEnp4m8D3mKby",
    "704 12th st west des moines ia": "https://drive.google.com/drive/folders/1gvSR2NBZLZVh4nubVCB7xTwCoSfbOzAj",
    "125 nw aurora ave des moines ia": "https://drive.google.com/drive/folders/1l4keF6hci04lNFz97_i7p6AzZDyYGImA",
    "4784 ne 22nd st - des moines ia": "https://drive.google.com/drive/folders/185DWXvII-kJhH2trg2LqLCPLf0uOlPXi",
    "125 nw aurora ave - des moines ia": "https://drive.google.com/drive/folders/1l4keF6hci04lNFz97_i7p6AzZDyYGImA",
    "prairie city duplexes portfolio - prairie city": "https://drive.google.com/drive/folders/1Z631_AslLr-x8-RZ3wuqnx2QhkMdm-Zk",
    "2101 university ave - des moines ia": "https://drive.google.com/drive/folders/19Vc23ml4mLQ_HZGjxKDdK7i6eGKduYZs",
    "2220 e ovid ave - des moines ia": "https://drive.google.com/drive/folders/18wB9kwTR6T_qfi8OGWLMUjF9Aes4T66D",
    "149 s chestnut ave - earlham ia": "https://drive.google.com/drive/folders/1w1C2erF2OR3OoTcPnzfBMPmrj1DqrRCS",
    "4719 toronto st - ames ia": "https://drive.google.com/drive/folders/1mzMI9eQYkPVT4vWzRNt97KdnqTVqUA8m",
    "420 water street - center point": "https://drive.google.com/drive/folders/1Qz7mCD5ats6k3FU6V5b8KrsSiVpyH8Ka",
    "2705 sw 9th st - des moines ia 50315": "https://drive.google.com/drive/folders/1xB4wShskRCXFzi5Gngpluuuk7Nu-xpin",
    "2705 sw 9th st des moines ia 50315": "https://drive.google.com/drive/folders/1xB4wShskRCXFzi5Gngpluuuk7Nu-xpin",
    "909 e 17th st - norwalk ia": "https://drive.google.com/drive/folders/1N66hV8XG_4REXMextrHdck4nzXdRhDyF",
    "315 e mckinley ave - des moines ia": "https://drive.google.com/drive/folders/1rhQpmYT0WXUOlrN_Rg1qwbMAqTi-7-Uq",
    "200 n 1st st - carlisle ia": "https://drive.google.com/drive/folders/1oc-YCHqR5p-ldo2PPVsP49d2_r85-GYq",
    "2411 welbeck road - des moines ia": "https://drive.google.com/drive/folders/1eVjfVu9C5HH4sJpidlT71-ANa-KVHfXj",
    "1416 3rd ave se - cedar rapids ia": "https://drive.google.com/drive/folders/1DyBtZMPT6BUlcBHMcFmQ_PwY1CD6yl1n",
    "200 e arlington st waterloo ia": "https://drive.google.com/drive/folders/1HUHX_ITF0oh0ZiaaG0hiym1_scVE0-6L",
    "2020 se 44th ct grimes ia": "https://drive.google.com/drive/folders/1ITyvQWY77-QihaWJLz6lJB8EKi_mQDpJ",
    "600 8th st se altoona ia": "https://drive.google.com/drive/folders/1_p8EKFsH-AAAksg3yjztfgWSUl7w23bu",
    "2411 welbeck rd des moines ia": "https://drive.google.com/drive/folders/1eVjfVu9C5HH4sJpidlT71-ANa-KVHfXj",
    "1513 n ankeny blvd ankeny ia": "https://drive.google.com/drive/folders/1B1uaQqFFMsFyq1xtbLUOEtNlJWr8I_PO",
    "104 arden st waterloo ia": "https://drive.google.com/drive/folders/1UcviwkwTq0uvuNWi1DPoUjkAKMWN28CJ",
    "405 n davis st anamosa ia": "https://drive.google.com/drive/folders/1Xxif6EUUvNgsWwGVRFKvLF1oUKLn2gyw",
    "1805 arlington ave - des moines ia": "https://drive.google.com/drive/folders/1n84bPFUH_Ou2-2ootIorkijb6IIqFdsQ",
    "1137 26th st des moines ia": "https://drive.google.com/drive/folders/1eY1cT2Sf2TQQeud2RfC6pCm6RywK5HKT",
    "3821 juniper trace unit 203 austin tx": "https://drive.google.com/drive/folders/1oNEvu5i1I2EdXVEW2r-I151EkSutfATy",
    "2733-2897 boston ave des moines ia": "https://drive.google.com/drive/folders/1fYNahVQ8MjU_bK3-RTjkqL8qY0wql0xX"
  }
} as const;